% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot_cont}
\alias{calibration_plot_cont}
\title{Create a calibration plot following meta-analysis THIS IS NOT OPTIMAL}
\usage{
calibration_plot_cont(predictions_df, study_var_name)
}
\value{
Creates a plot showing all predicted and observed values from  internal-external cross-validation as well as the calibraion slope line fom the meta-analysis.
}
\description{
This may not be the optimal approach as smoothed calibraiton is plotted using all observations from internal-external cross-validation.
}
