% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formaters.R
\name{compress_columns}
\alias{compress_columns}
\title{Format results data.frame so it has text columns with intervals in}
\usage{
compress_columns(results_df, digits = 2)
}
\arguments{
\item{results_df}{A data frame with meta-analysis results}

\item{digits}{The number of decimal places to use with prediciton and confidence intervals}
}
\value{
A formatted data.frame
}
\description{
Format results data.frame so it has text columns with intervals in
}
