% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bayes_predictors.R
\name{eb}
\alias{eb}
\title{Empirical Bayes predictions for 2 level mixed effects models estimated using lme4 THIS FUNCTION DOES NOT WORK}
\usage{
eb(model, newdata)
}
\arguments{
\item{model}{An lme4 model object}

\item{newdata}{A data frame with the same structure as the data used to fit the model}
}
\value{
Returns a data.frame, the first column is the cluster variable and the second column is the empirical bayes prediction for the intercept.
}
\description{
Empirical Bayes predictions for 2 level mixed effects models estimated using lme4 THIS FUNCTION DOES NOT WORK
}
