% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{evaluate_performance_cont_obs_pred}
\alias{evaluate_performance_cont_obs_pred}
\title{Calculates model performance metrics for a continuous outcome vectors of predicted values and actuals}
\usage{
evaluate_performance_cont_obs_pred(actual, predicted)
}
\arguments{
\item{A}{vector of predicted outcomes}
}
\value{
A data.frame with columns for metric name, estimate and standard error.
}
\description{
Calculates model performance metrics for a continuous outcome vectors of predicted values and actuals
}
