% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{evaluate_performance_cont_obs_pred_calib_slope_int}
\alias{evaluate_performance_cont_obs_pred_calib_slope_int}
\title{Calibration slope and intercept for a continuous outcome}
\usage{
evaluate_performance_cont_obs_pred_calib_slope_int(actual, predicted)
}
\value{
A data.frame with columns for metric name, estimate and standard error.
}
\description{
Calibration slope and intercept for a continuous outcome
}
