% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma_prediction_pipeline.R
\name{get_performance_by_study}
\alias{get_performance_by_study}
\title{Title}
\usage{
get_performance_by_study(
  by_study_predictions_df,
  evaluate_performance,
  study_var_name
)
}
\arguments{
\item{by_study_predictions_df}{A dataframe containing individaul level predictions in the 'pred' column, observed values in the 'actual' column and a column indcating study}

\item{evaluate_performance}{The function used to evaluate performance}

\item{study_var_name}{The name of the study column}
}
\value{
A data.frame containing the performance of the model for each study
}
\description{
Title
}
