% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bayes_predictors.R
\name{get_rand_int}
\alias{get_rand_int}
\title{Returns empirical bayes predictions for random intercepts for new clusters for a random intercept model estimated with lme4}
\usage{
get_rand_int(model, newdata)
}
\arguments{
\item{model}{An lme4 model object}

\item{newdata}{A data frame with the same structure as the data used to fit the model}
}
\value{
Returns a data.frame, the first column is the cluster variable and the second column is the empirical bayes prediction for the intercept.
}
\description{
Returns empirical bayes predictions for random intercepts for new clusters for a random intercept model estimated with lme4
}
