% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma_prediction_pipeline.R
\name{ipdma_prediction_pipeline}
\alias{ipdma_prediction_pipeline}
\title{Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.}
\usage{
ipdma_prediction_pipeline(
  data,
  model_function,
  InternalExternalCV = TRUE,
  predict_function = predict,
  evaluate_performance,
  test_data = NULL,
  out_var_name,
  study_var_name = "studyid"
)
}
\arguments{
\item{evaluate_performance}{a function that evaluates the performance of the model for a single study}

\item{test_data}{data used to validate the model}

\item{train_data}{data used to fit the model}

\item{fit_model}{a function that fits the model to be used}
}
\description{
Pipeline for developing and validating a prediction model by study for an IPD Meta-analysis with training and test data.
}
