% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDPredictR-package.R
\docType{package}
\name{IPDPredictR-package}
\alias{IPDPredictR}
\alias{IPDPredictR-package}
\title{IPDPredictR: Funcitons to aid with the development and validaion of preidciton mdoels using IPD meta-analysis.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
