% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdma_prediction_pipeline.R
\name{meta_analyse_predictions_cont}
\alias{meta_analyse_predictions_cont}
\title{Meta-anlyse predictions of a continuous outcome}
\usage{
meta_analyse_predictions_cont(predictions, study_var_name)
}
\arguments{
\item{predictions}{A data frame containing individual level predictions in the 'pred' column, observed values in the 'actual' column and a column indicating study}

\item{study_var_name}{A string containing the name of the column indicating study}
}
\value{
The results of the meta analysis
}
\description{
Meta-anlyse predictions of a continuous outcome
}
