% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analyse_predictions_cont_mi.R
\name{meta_analyse_predictions_cont_mi}
\alias{meta_analyse_predictions_cont_mi}
\title{Meta-analyse predictions for continuous outcomes with multiple imputation}
\usage{
meta_analyse_predictions_cont_mi(
  predictions,
  study_var_name,
  imp_indicator_name
)
}
\arguments{
\item{predictions}{A data frame containing individual level predictions in the 'pred' column, observed values in the 'actual' column and a column indicating study}

\item{study_var_name}{A string containing the name of the column indicating study}

\item{imp_indicator_name}{The name of the column indicating which imputed data set was used}
}
\value{
The results of the meta analysis
}
\description{
Meta-analyse predictions for continuous outcomes with multiple imputation
}
