% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{predict_fixed}
\alias{predict_fixed}
\title{Predictions from a model excluding the cluster variable}
\usage{
predict_fixed(model, newdata)
}
\arguments{
\item{model}{A model, either lm or lmerMod}

\item{newdata}{A data frame with the same structure as the data used to fit the model}
}
\value{
A vector of predicted values
}
\description{
Provides predictions for linear predictor of a model excluding the study var for random intercept or linear regression models.
For random intercept models it returns the predictions with random intercepts set to zero.
For linear regression models it returns the predictions based only on variables that start with the letter "x"
}
