% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{predict_intercepts}
\alias{predict_intercepts}
\title{Prediction of cluster intercepts}
\usage{
predict_intercepts(model, newdata, cluster_var = "studyid")
}
\arguments{
\item{model}{A model, either lm or lmerMod}

\item{newdata}{A data frame with the same structure as the data used to fit the model}

\item{cluster_var}{The name of the cluster variable as a string. Note for lm models it assumes other predictors start with "x"}
}
\value{
A data.frame with the first column being the cluster variable and the second column being the predicted intercept.
}
\description{
For lmerMod models it returns empirical Bayes predicitons of intercepts. For lm models it returns maximum liklihood estaitmes.
}
